refactor: Quitar componentes no utilizados, actualizar BenefitList. #244
