--- src/teacher/components/benefitsView/benefitPurchases/benefitPurchaseHeader/BenefitPurchasesHeader.module.css
@@ -1,88 +1,0 @@
-@import "../../../../../shared/styles/colors.css";
-
-.headerContainer {
-  margin-bottom: 2rem;
-}
-
-.headerTop {
-  display: grid;
-  grid-template-columns: auto 1fr auto;
-  align-items: center;
-  gap: 1.5rem;
-}
-
-.backButton {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.5rem;
-  font-weight: 500;
-  white-space: nowrap;
-}
-
-.headerContent {
-  flex: 1;
-}
-
-.title {
-  font-size: 2.5rem;
-  font-weight: bold;
-  color: var(--color-text-title);
-  margin: 0 0 0.5rem 0;
-  background: linear-gradient(
-    145deg,
-    var(--color-primary) 40%,
-    var(--color-third)
-  );
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  background-clip: text;
-}
-
-.subtitle {
-  color: var(--color-text-subtitle);
-  font-size: 1.125rem;
-  line-height: 1.6;
-  margin: 0;
-}
-
-.toggleButton {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.5rem;
-  font-weight: 500;
-  white-space: nowrap;
-}
-
-/* Responsive Design */
-@media (max-width: 1024px) {
-  .headerTop {
-    grid-template-columns: 1fr;
-    gap: 1rem;
-  }
-
-  .backButton {
-    justify-self: flex-start;
-  }
-
-  .toggleButton {
-    justify-self: flex-start;
-  }
-
-  .title {
-    font-size: 2rem;
-  }
-
-  .subtitle {
-    font-size: 1rem;
-  }
-}
-
-@media (max-width: 768px) {
-  .title {
-    font-size: 1.75rem;
-  }
-
-  .subtitle {
-    font-size: 0.9375rem;
-  }
-}
